+++ NEWS.md
@@ -1,7 +1,7 @@
 # Version 0.2.32
 
 * allow choice of storage format in `s3read` and `s3store`. Defaults to `RDS`,
-  and now you can chooose `CSV` for data frames.
+  and now you can chooose `CSV` or `table` for data frames.
 
 # Version 0.2.31